+++ src/main.js
@@ -16,7 +16,7 @@ import { handleError } from './error-handler';
 
 // load without importing
 import { } from './ui/icons/git-icon';
-import { } from './ui/branch';
+import { } from './lib/branch';
 import { } from './ui/panel';
 import { } from './menu-entries';
