+++ src/ui/branch.js
@@ -1,74 +1,31 @@
-import Events from '../events';
-import EventEmitter from '../event-emitter';
-import Preferences from '../preferences';
-import Strings from 'strings';
-import getRoot from '../git/get-root';
-import getBranchName from '../git/get-branch-name';
-import getMergeInfo from '../git/get-merge-info';
-import { _ } from '../brackets';
-import { getProjectRoot } from '../utils';
-import { handleError } from '../error-handler';
-
 const $branchContainer = $(`<div id="brackets-git-branch" class="btn-alt-quiet"/>`);
 const $branchName = $(`<span class="branch-name">\u2026</span>`);
-const MAX_LEN = 20;
 
 function toggleDropdown() {
 
 }
 
-async function refresh() {
-  $branchName.text('\u2026').parent().show();
-
-  let gitRoot = await getRoot();
-  let projectRoot = getProjectRoot();
-  let isRepositoryRootOrChild = gitRoot && projectRoot.startsWith(gitRoot);
-
-  $branchName.parent().toggle(isRepositoryRootOrChild);
-
-  if (!isRepositoryRootOrChild) {
-    Preferences.set('currentGitRoot', projectRoot);
-    Preferences.set('currentGitSubfolder', '');
-    EventEmitter.emit(Events.GIT_REPO_NOT_AVAILABLE, Strings.PROJECT_NOT_A_REPO);
-    return;
-  }
-
-  Preferences.set('currentGitRoot', gitRoot);
-  Preferences.set('currentGitSubfolder', projectRoot.substring(gitRoot.length));
-
-  let fullBranchName = await getBranchName();
-  let branchNameTooLong = fullBranchName.length > MAX_LEN;
-  let branchName = branchNameTooLong ? fullBranchName.substring(0, MAX_LEN) + '\u2026' : fullBranchName;
-
-  let mergeInfo = await getMergeInfo();
-
-  if (mergeInfo.mergeMode) {
-    branchName += '|MERGING';
-  }
-
-  if (mergeInfo.rebaseMode) {
-    if (mergeInfo.rebaseHead) {
-      branchName = mergeInfo.rebaseHead;
-    }
-    branchName += '|REBASE';
-    if (mergeInfo.rebaseNext && mergeInfo.rebaseLast) {
-      branchName += '(' + mergeInfo.rebaseNext + '/' + mergeInfo.rebaseLast + ')';
-    }
-  }
-
+export function toggle(bool) {
   $branchName
-    .text(branchName)
-    .attr('title', branchNameTooLong ? fullBranchName : null)
+    .parent()
+    .toggle(bool);
+}
+
+export function setText(text) {
+  $branchName.text(text);
+  toggle(true);
+}
+
+export function setBranchName(text, title) {
+  setText(text);
+  $branchName
+    .attr('title', title)
     .off('click')
     .on('click', toggleDropdown)
     .append($('<span class="dropdown-arrow" />'));
-
-  EventEmitter.emit(Events.GIT_REPO_AVAILABLE);
-  EventEmitter.emit(Events.REBASE_MERGE_MODE, mergeInfo.rebaseMode, mergeInfo.mergeMode);
 }
 
-const init = _.once(function () {
-
+export function init() {
   $branchContainer
     .append('[ ')
     .append($branchName)
@@ -78,11 +35,4 @@ const init = _.once(function () {
       return false;
     })
     .appendTo('#project-files-header');
-
-  refresh().catch(err => handleError(err));
-
-  // TODO: attach events triggering refresh here
-
-});
-
-EventEmitter.on(Events.GIT_WORKING, () => init());
+}
